requirements.txt
Django==2.2.1
requests==2.21.0
urllib3==1.24.1
djangorestframework==3.10.2
redis==3.3.11
